+++ pyzoo/zoo/tfpark/text/__init__.py
@@ -14,6 +14,5 @@
 # limitations under the License.
 #
 
-from zoo.util.tf import variable_creator_scope
-from zoo.pipeline.api.net import TFDataset
-from zoo.tfpark.model import KerasModel
+from .text_model import *
+from .intent_extraction import *
